requirements.txt
streamlit
yfinance
pandas
matplotlib
python-dotenv
langchain-core
langchain-openai